move BuildingBuilderView to a new package buildingviews more sonar fixes
