+++ src/main/java/com/minecolonies/coremod/colony/buildings/buildingviews/BuildingBuilderView.java
@@ -1,4 +1,4 @@
-package com.minecolonies.coremod.colony.buildings;
+package com.minecolonies.coremod.colony.buildings.buildingviews;
 
 import com.minecolonies.blockout.views.Window;
 import com.minecolonies.coremod.client.gui.WindowHutBuilder;
